Fix issue #292 - Metadata size not accounted for in ChunkStore size chunking
When ChunkStore chunks for size (due to 16MB doc limit in MongoDB), it was not including the size of the metadata in the size of the document. In general, this is ok, the metadata is very small, but it can be quite large when array masks are included in the metadata (this happens when there are columns of strings that have missing values).