Fix read of empty MultiIndex+tz Series
It's impossible to determine the dtype of a saved empty index from the saved information alone and pandas has the bad habit of defaulting to either float or int. We can work out whether a MultIndex level is meant to be a DatetimeIndex based on the presence of the separate timezone information. Based on this, we can cast back empty indices to DatetimeIndex where appropriate.

This is only a shim - a good fix would have us saving down all the dtypes for both indices and data. This should ensure that any arbitrary type empty series written out can be read back in as it was.