Add a cache option for list_libraries
list_libraries is a frequently called operation that can get
quite slow with a high amount of libraries. I noticed 4-5s+
for 7-8k libraries. Also the number of libraries is fairly
constant, so not much point in refetching the entire list every
time.

This adds a new collection in mongo which acts as a global cache
for all queries. The meta_db can be used for more accounting and
metadata storage in future, and this collection can be used for
caching other stuff as well.