Fix #587 Add a best effort cache option for list_libraries
list_libraries is a frequently called operation that can get
quite slow with a high amount of libraries. I noticed 4-5s+
for 7-8k libraries. Also the number of libraries is fairly
constant, so not much point in refetching the entire list every
time.

This adds a new collection in mongo which acts as a global cache
for all queries. The meta_db can be used for more accounting and
metadata storage in future, and this collection can be used for
caching other stuff as well.

It's best effort as if you don't have permissions or the data
does not exist in the cache it will just use the list_libraries.
The idea is to not require admin for reads and not expect the user
to have write access to a collection on writes.